Number Base System Translator
Here is my final version of number base translator from 2020. For example, it can translate numbers in binary to numbers in decimal or ternary. It supports bases from 1 to 62. 

It first uses the regular decimal digits to display the numbers. When you go to bases over 10 the code will start to use uppercase letters from English alphabet as digits and after base 36 the code will start to use lowercase letters from English alphabet as digits.

There are two versions: 
- "nj6_5.py" This is the original and it's in Finnish. "nj" comes from "numero järjestelmä" which means number system in Finnish. The numbers "6_5" means the iteration of the code. So it's the 6th major iteration and 5th small iteration of the 6th major iteration.
- "nj6_5_en.py" is the same code but I've translated it to English.